// src/app/master/models/esanchit.ts
import { GlobalVariables } from '../../core/models/globalvariables';

export class Esanchit {
  doc_pkid: string;
  doc_drn: string;
  doc_irn: string;
  doc_type_code: string;

  doc_file_type: string;
  doc_file_name: string;
  doc_name: string;
  doc_ref_no: string;
  doc_upload_date: string;

  doc_issued_at: string;
  doc_issued_date: string;
  doc_expiry_date: string;

  doc_issuer_code: string;
  doc_issuer_name: string;
  doc_issuer_add1: string;
  doc_issuer_add2: string;
  doc_issuer_city: string;
  doc_issuer_pin: string;

  doc_ben_code: string;
  doc_ben_name: string;
  doc_ben_add1: string;
  doc_ben_add2: string;
  doc_ben_city: string;
  doc_ben_pin: string;
  doc_selected: boolean;

  rec_company_code: string;
  rec_branch_code: string;
  rec_category: string;
  doc_job_year: string;
  doc_job_id: string;
  doc_link_id: string;
  doc_link_type: string;

  doc_email_slno: number;
  rec_mode: string;
  
  doc_job_no: string;      
  doc_type_name: string;
  doc_all_drn_selected: boolean;

  _globalvariables: GlobalVariables;

}